Refactored some function names.
